Made DTD's internal, interlinked documents using id's
For all XML files I have made all DTD's internal. While doing this I noticed that we didn't have to retype redundant data, we could just reference ID's in master databases for visitors and showpieces and then in an XQuery, could cross reference the id variables to get relevant information. So I made exhibition a master database that has ID's for visitors and showpieces, so that we can make an XQuery to get relevant information for each exhibition (will commit the xquery next)